TMS/requirements.txt
PyQt5==5.15.10
fastapi==0.104.1
uvicorn[standard]==0.24.0
mysql-connector-python==8.2.0
bcrypt==4.1.1
cryptography==41.0.7
pyinstaller==6.2.0
python-dotenv==1.0.0
email-validator==2.1.0
pydantic==2.5.0
requests==2.31.0
websockets==12.0